UserIdGenerator implementation.
