+++ sdk/src/main/java/com/qonversion/android/sdk/internal/user/generator/UserIdGenerator.kt
@@ -1,0 +1,6 @@
+package com.qonversion.android.sdk.internal.user.generator
+
+interface UserIdGenerator {
+
+    fun generateUserId(): String
+}
